+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV008 - 28/03/2003 - Add the EventInformation class to get information of event.
 - LASRV008 - 28/03/2003 - Add the EventInsertion class to insert event.
 - LASRV008 - 28/03/2003 - Add the EventModificator class to modify event.
+- LASRV008 - 28/03/2003 - Add the EventDeletion class to delete event.
 
 
 ### Changed